I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

Spit `right_table`, and make tasks number close to thread number. ~5x 

TPCH SF 0.01
Before
```
mysql> select count() from lineitem join orders on l_orderkey > o_orderkey and l_partkey < o_custkey;
+-----------+
| count()   |
+-----------+
| 169757135 |
+-----------+
1 row in set (7.55 sec)
Read 75175 rows, 1.15 MiB in 7.526 sec., 9.99 thousand rows/sec., 156.08 KiB/sec.
```

Now
```
mysql> select count() from lineitem join orders on l_orderkey > o_orderkey and l_partkey < o_custkey;
+-----------+
| count()   |
+-----------+
| 169757135 |
+-----------+
1 row in set (1.63 sec)
Read 75175 rows, 1.15 MiB in 1.601 sec., 46.96 thousand rows/sec., 733.67 KiB/sec.
```

- Closes #issue

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/13229)
<!-- Reviewable:end -->